+++ packages/agtree/src/parser/cosmetic/html-filtering-body/ubo-html-filtering-body-parser.ts
@@ -95,7 +95,7 @@ export class UboHtmlFilteringBodyParser extends BaseParser {
         baseOffset = 0,
     ): HtmlFilteringRuleBody | null {
         // If HTML filtering rules parsing is disabled, return null
-        if (!options.parseHtmlFilteringRules) {
+        if (!options.parseHtmlFilteringRuleBodies) {
             return null;
         }