+++ packages/agtree/test/parser/cosmetic/html-filtering-body/adg-html-filtering.test.ts
@@ -21,7 +21,7 @@ import { defaultParserOptions, type ParserOptions } from '../../../../src/parser
  */
 const parsingEnabledDefaultParserOptions: ParserOptions = {
     ...defaultParserOptions,
-    parseHtmlFilteringRules: true,
+    parseHtmlFilteringRuleBodies: true,
 };
 
 /**