+++ packages/agtree/test/parser/cosmetic/html-filtering-body/html-filtering.test.ts
@@ -27,7 +27,7 @@ import { defaultParserOptions, type ParserOptions } from '../../../../src/parser
  */
 const parsingEnabledDefaultParserOptions: ParserOptions = {
     ...defaultParserOptions,
-    parseHtmlFilteringRules: true,
+    parseHtmlFilteringRuleBodies: true,
 };
 
 /**