+++ packages/agtree/test/parser/cosmetic/html-filtering-body/ubo-html-filtering.test.ts
@@ -26,7 +26,7 @@ import { defaultParserOptions, type ParserOptions } from '../../../../src/parser
  */
 const parsingEnabledDefaultParserOptions: ParserOptions = {
     ...defaultParserOptions,
-    parseHtmlFilteringRules: true,
+    parseHtmlFilteringRuleBodies: true,
 };
 
 /**